add failsafe profile for running docker based integration tests
