+++ edc-controlplane/edc-controlplane-memory/pom.xml
@@ -234,7 +234,6 @@
         </dependency>
     </dependencies>
 
-
     <profiles>
         <profile>
             <id>with-docker-image</id>
