+++ edc-extensions/hashicorp-vault/src/test/java/org/eclipse/dataspaceconnector/core/security/hashicorp/HashicorpCertificateResolverIT.java
@@ -14,9 +14,6 @@
 
 package org.eclipse.dataspaceconnector.core.security.hashicorp;
 
-import static org.eclipse.dataspaceconnector.core.security.hashicorp.HashicorpVaultExtension.VAULT_TOKEN;
-import static org.eclipse.dataspaceconnector.core.security.hashicorp.HashicorpVaultExtension.VAULT_URL;
-
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
@@ -25,7 +22,6 @@ import java.nio.charset.StandardCharsets;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
 import java.security.Provider;
 import java.security.PublicKey;
 import java.security.SecureRandom;
@@ -34,10 +30,8 @@ import java.security.cert.X509Certificate;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Optional;
 import java.util.UUID;
-import lombok.Getter;
 import lombok.SneakyThrows;
 import org.bouncycastle.asn1.oiw.OIWObjectIdentifiers;
 import org.bouncycastle.asn1.x500.X500Name;
@@ -57,52 +51,17 @@ import org.bouncycastle.operator.DigestCalculator;
 import org.bouncycastle.operator.OperatorCreationException;
 import org.bouncycastle.operator.bc.BcDigestCalculatorProvider;
 import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
-import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
 import org.eclipse.dataspaceconnector.spi.security.CertificateResolver;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.junit.ClassRule;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.shaded.org.bouncycastle.openssl.jcajce.JcaPEMWriter;
-import org.testcontainers.utility.DockerImageName;
-import org.testcontainers.vault.VaultContainer;
 
-@Testcontainers
-@ExtendWith(EdcExtension.class)
-class HashicorpCertificateResolverTest {
-  private static final String DOCKER_IMAGE_NAME = "vault:1.9.6";
-  private static final String TEST_TOKEN = "test-token";
+class HashicorpCertificateResolverIT extends AbstractHashicorpIT {
   private static final Provider PROVIDER = new BouncyCastleProvider();
   private static final JcaX509CertificateConverter JCA_X509_CERTIFICATE_CONVERTER =
       new JcaX509CertificateConverter().setProvider(PROVIDER);
   private static final String SIGNATURE_ALGORITHM = "SHA256WithRSAEncryption";
-
-  private final TestExtension testExtension = new TestExtension();
-
-  @Container @ClassRule
-  private static final VaultContainer<?> vaultContainer =
-      new VaultContainer<>(DockerImageName.parse(DOCKER_IMAGE_NAME)).withVaultToken(TEST_TOKEN);
-
-  @BeforeEach
-  void beforeEach(EdcExtension extension) {
-    extension.setConfiguration(
-        new HashMap<>() {
-          {
-            put(
-                VAULT_URL,
-                String.format(
-                    "http://%s:%s", vaultContainer.getHost(), vaultContainer.getFirstMappedPort()));
-            put(VAULT_TOKEN, TEST_TOKEN);
-          }
-        });
-    extension.registerSystemExtension(ServiceExtension.class, testExtension);
-  }
 
   @Test
   @SneakyThrows
@@ -111,9 +70,9 @@ class HashicorpCertificateResolverTest {
     X509Certificate certificateExpected = generateCertificate(5, "Test");
     String pem = convertToPem(certificateExpected);
 
-    Vault vault = testExtension.getVault();
+    Vault vault = getVault();
     vault.storeSecret(key, pem);
-    CertificateResolver resolver = testExtension.getCertificateResolver();
+    CertificateResolver resolver = getCertificateResolver();
     X509Certificate certificateResult = resolver.resolveCertificate(key);
 
     Assertions.assertEquals(certificateExpected, certificateResult);
@@ -124,17 +83,17 @@ class HashicorpCertificateResolverTest {
   void resolveCertificate_malformed() {
     String key = UUID.randomUUID().toString();
     String value = UUID.randomUUID().toString();
-    Vault vault = testExtension.getVault();
+    Vault vault = getVault();
     vault.storeSecret(key, value);
 
-    CertificateResolver resolver = testExtension.getCertificateResolver();
+    CertificateResolver resolver = getCertificateResolver();
     X509Certificate certificateResult = resolver.resolveCertificate(key);
     Assertions.assertNull(certificateResult);
   }
 
   private static X509Certificate generateCertificate(int validity, String cn)
-      throws CertificateException, OperatorCreationException, IOException, NoSuchAlgorithmException,
-          NoSuchProviderException {
+      throws CertificateException, OperatorCreationException, IOException,
+          NoSuchAlgorithmException {
 
     KeyPair keyPair = generateKeyPair();
 
@@ -167,8 +126,7 @@ class HashicorpCertificateResolverTest {
     return JCA_X509_CERTIFICATE_CONVERTER.getCertificate(certificateBuilder.build(contentSigner));
   }
 
-  private static KeyPair generateKeyPair()
-      throws NoSuchAlgorithmException, NoSuchProviderException {
+  private static KeyPair generateKeyPair() throws NoSuchAlgorithmException {
     KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA", PROVIDER);
     keyPairGenerator.initialize(1024, new SecureRandom());
 
@@ -200,18 +158,6 @@ class HashicorpCertificateResolverTest {
         pemWriter.flush();
       }
       return stream.toString(StandardCharsets.UTF_8);
-    }
-  }
-
-  @Getter
-  private static class TestExtension implements ServiceExtension {
-    private Vault vault;
-    private CertificateResolver certificateResolver;
-
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-      vault = context.getService(Vault.class);
-      certificateResolver = context.getService(CertificateResolver.class);
     }
   }
 }
